+++ src/main/java/com/example/eduspringBootfinal/repository/DetalleOrdenRepository.java
@@ -3,5 +3,5 @@ package com.example.eduspringBootfinal.repository;
 import com.example.eduspringBootfinal.entities.DetalleOrden;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetalleOrdenRepository extends JpaRepository<DetalleOrden, Long> {
+public interface DetalleOrdenRepository extends JpaRepository<DetalleOrden, Integer> {
 }
