+++ src/main/java/com/example/eduspringBootfinal/repository/OrdenRepository.java
@@ -3,5 +3,5 @@ package com.example.eduspringBootfinal.repository;
 import com.example.eduspringBootfinal.entities.Orden;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdenRepository extends JpaRepository<Orden, Long> {
+public interface OrdenRepository extends JpaRepository<Orden, Integer> {
 }
